+++ R/echarty.R
@@ -6,7 +6,7 @@
 #'
 #' @param df A data.frame to be preset as \href{https://echarts.apache.org/en/option.html#dataset}{dataset}, default NULL \cr
 #'     For crosstalk df should be of type \code{\link[crosstalk]{SharedData}}.
-#' @param group1 Type of grouped series, default 'scatter'. Set to NULL to disable. \cr
+#' @param group1 Type of grouped series, or type of first ungrouped serie. Default is 'scatter'. Set to NULL to disable. \cr
 #'     If the grouping is on multiple columns, only the first one is used.
 #' @param preset Disable(FALSE) or enable (TRUE, default) presets xAxis,yAxis,serie for 2D, or grid3D,xAxis3D,yAxis3D,zAxis3D for 3D.
 #' @param load Name(s) of plugin(s) to load. Could be a character vector or comma-delimited string. default NULL.\cr
@@ -54,8 +54,9 @@ ec.init <- function( df = NULL, group1 = 'scatter', preset = TRUE, load = NULL,
   if (preset) {
     if (!('xAxis' %in% names(opts))) opts$xAxis <- list(ey='')
     if (!('yAxis' %in% names(opts))) opts$yAxis <- list(ey='')
-    if (!('series' %in% names(opts))) opts$series <- list(list())
-    opts$series[[1]] <- list(type='scatter')
+    if (!('series' %in% names(opts))) opts$series <- list(
+    	list(type=if (is.null(group1)) 'scatter' else group1) )
+    #opts$series[[1]] <- list(type='scatter')
   }
 
   
@@ -100,10 +101,10 @@ ec.init <- function( df = NULL, group1 = 'scatter', preset = TRUE, load = NULL,
       x$opts$legend = list(data=list())
       for(i in grvals) { 
         k <- k+1
-        srch4 <- i
-        if ('factor' %in% class(grvals)) srch4 <- k
+        #srch4 <- i
+        #if ('factor' %in% class(grvals)) srch4 <- k
         txfm <- append(txfm, list(list(transform = list(
-          type='filter', config=list(dimension=grnm, '='=srch4)))))
+          type='filter', config=list(dimension=grnm, '='=i)))))
         x$opts$series[[k]] <- list(
           type=group1, datasetIndex=k, name=as.character(i))
         x$opts$legend$data <- append(x$opts$legend$data, list(list(name=as.character(i))))
@@ -189,7 +190,7 @@ ec.init <- function( df = NULL, group1 = 'scatter', preset = TRUE, load = NULL,
       wt$x$opts$yAxis3D <- list(list())
       wt$x$opts$zAxis3D <- list(list())
     }
-    wt <- ec.plugjs(wt, 'https://cdn.jsdelivr.net/npm/echarts-gl@2.0.2/dist/echarts-gl.min.js')
+    wt <- ec.plugjs(wt, 'https://cdn.jsdelivr.net/npm/echarts-gl@2.0.4/dist/echarts-gl.min.js')
   }
   if ('world' %in% load) 
     wt <- ec.plugjs(wt, 'https://cdn.jsdelivr.net/npm/echarts@4.9.0/map/js/world.js')
@@ -198,7 +199,7 @@ ec.init <- function( df = NULL, group1 = 'scatter', preset = TRUE, load = NULL,
     wt <- ec.plugjs(wt, 'https://cdn.jsdelivr.net/npm/echarts-liquidfill@3.0.0/dist/echarts-liquidfill.min.js')
   
   if ('gmodular' %in% load) 
-    wt <- ec.plugjs(wt, 'https://github.com/ecomfe/echarts-graph-modularity/raw/master/dist/echarts-graph-modularity.min.js')
+    wt <- ec.plugjs(wt, 'https://cdn.jsdelivr.net/npm/echarts-graph-modularity@2.0.0/dist/echarts-graph-modularity.min.js')
   
   if ('wordcloud' %in% load) 
     wt <- ec.plugjs(wt, 'https://cdn.jsdelivr.net/npm/echarts-wordcloud@2.0.0/dist/echarts-wordcloud.min.js')
@@ -275,15 +276,15 @@ ec.plugjs <- function(wt=NULL, source=NULL) {
 }
 
 
-#' ECharts data helper
+#' Data helper
 #' 
 #' Make ECharts data from a data.frame
 #' 
-#' @param df Chart data in data.frame format, required
+#' @param df Chart data in data.frame format, required. 
 #' @param format A key on how to format the output list \cr \itemize{
-#'  \item 'dataset' = list used in \href{https://echarts.apache.org/en/option.html#dataset.source}{dataset} (default), or in \href{https://echarts.apache.org/en/option.html#series-scatter.data}{series.data} but without a header.\cr
+#'  \item 'dataset' = list to be used in \href{https://echarts.apache.org/en/option.html#dataset.source}{dataset} (default), or in \href{https://echarts.apache.org/en/option.html#series-scatter.data}{series.data} but without a header. \cr
 #'  \item 'values' = list for customized \href{https://echarts.apache.org/en/option.html#series-scatter.data}{series.data} \cr
-#'  \item 'names' = named lists useful for named data like \href{https://echarts.apache.org/en/option.html#series-sankey.links}{sankey links}
+#'  \item 'names' = named lists useful for named data like \href{https://echarts.apache.org/en/option.html#series-sankey.links}{sankey links}.
 #'  }
 #' @param header Whether the data will have a header with column names or not, default TRUE. Set this to FALSE when used in \href{https://echarts.apache.org/en/option.html#series-scatter.data}{series.data}.
 #' @return A list for \emph{dataset.source}, \emph{series.data} or a list of named lists.
@@ -296,7 +297,14 @@ ec.data <- function(df, format='dataset', header=TRUE) {
     stop('df has to be data.frame', call. = FALSE)
   
   # TODO: replace purrr with something simpler
-  tmp <- purrr::transpose(df)       # named lists
+  #tmp <- purrr::transpose(df)       # named lists
+  rownames(df) <- NULL
+  tmp <- apply(df, 1, function(x) {
+    out <- list()
+    i <- 1
+    for(n in colnames(df)) { out[n] <- x[i]; i <- i+1 }
+    out 
+  })
   if (format=='dataset') {
     datset <- lapply(tmp, unname)
     if (header)
@@ -371,6 +379,67 @@ ec.timegrp <- function(wt, df=NULL, scol=NULL, xcol=NULL, type='line', ...) {
   wt
 }
 
+
+
+#' Charts layout
+#' 
+#' Set multiple charts in rows/columns format
+#' 
+#' @param plots A list of charts
+#' @param rows Number of rows
+#' @param cols Number of columns
+#' @param width Width of columns, one of xs, md, lg
+#' @param title Title for the set
+#' @return A container \code{\link[htmltools]{div}} in rmarkdown, otherwise \code{\link[htmltools]{browsable}}
+#' @examples
+#' 
+#' p1 <- cars %>% ec.init()
+#' p2 <- cars %>% ec.init() %>% ec.theme('dark')
+#' ec.layout(list(p1,p2), cols=2 )
+#' 
+#' @export 
+ec.layout <- function (plots, rows = NULL, cols = NULL, width = "xs", 
+                       title = NULL) 
+{
+  if (!is.list(plots))
+    stop('ec.layout charts must be a list', call. = FALSE)
+  if (is.null(rows) & !is.null(cols)) rows <- ceiling(length(plots)/cols)
+  if (!is.null(rows) & is.null(cols)) cols <- ceiling(length(plots)/rows)
+  if (is.null(rows) & is.null(cols)) { rows <- length(plots); cols <- 1 }
+  w <- "-xs"
+  if (!isTRUE(getOption("knitr.in.progress"))) w <- ""
+  x <- 0
+  tg <- htmltools::tagList()
+  for (i in 1:rows) {
+    r <- htmltools::div(class = "row")
+    for (j in 1:cols) {
+      x <- x + 1
+      cl <- paste0("col", w, "-", 12/cols)
+      if (x <= length(plots))
+        c <- htmltools::div(class = cl, plots[[x]])
+      else 
+        c <- htmltools::div(class = cl)
+      r <- htmltools::tagAppendChild(r, c)
+    }
+    tg <- htmltools::tagAppendChild(tg, r)
+  }
+  if (isTRUE(getOption("knitr.in.progress"))) {
+    if (!is.null(title))
+      htmltools::div(title, tg)
+    else
+      tg
+  }
+  else
+    htmltools::browsable(
+      htmltools::div(
+        class = "container-fluid", 
+        htmltools::tags$head(
+          htmltools::tags$link(
+            rel = "stylesheet", 
+            href = "https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css"
+          )),
+        htmltools::h3(title), tg))
+}
 
 
 #' Area band
@@ -635,9 +704,10 @@ ecs.proxy <- function(id) {
 #'
 #' @param proxy A [ecs.proxy] object
 #' @param cmd Name of command, default is \emph{p_merge}\cr
-#'   Other proxy commands:\cr
+#'   The proxy commands are:\cr
 #' \emph{p_update} - add new series and axes\cr
 #' \emph{p_merge} - add serie features like marks\cr
+#' \emph{p_replace} - replace entire chart \cr
 #' \emph{p_del_serie} - delete a serie by index or name\cr
 #' \emph{p_del_marks} - delete marks of a serie\cr
 #' \emph{p_append_data} - add data to existing series\cr
@@ -670,9 +740,9 @@ ecs.proxy <- function(id) {
 #'   output$plot <- ecs.render({
 #'     p <- ec.init()
 #'     p$x$opts$series <- lapply(mtcars %>% relocate(disp, .after=mpg)
-#'                               %>% group_by(cyl) %>% group_split(), function(s) {
-#'                                 list(type='scatter', name=unique(s$cyl), data=ec.data(s, 'values'))
-#'                               })
+#'       %>% group_by(cyl) %>% group_split(), function(s) {
+#'           list(type='scatter', name=unique(s$cyl), data=ec.data(s, 'values'))
+#'       })
 #'     p$x$opts$legend <- list(ey='')
 #'     p$x$opts$xAxis <- list(type="value"); p$x$opts$yAxis <- list(ec='')
 #'     p$x$opts$tooltip <- list(list(show=TRUE))
@@ -719,7 +789,6 @@ ecs.proxy <- function(id) {
 #'   })
 #' 
 #'   observeEvent(input$adata, {
-#'     set.seed(sample(1:444, 1))
 #'     tmp <- apply(unname(data.frame(rnorm(5, 10, 3), rnorm(5, 200, 33))),
 #'                  1, function(x) { list(value=x) })
 #'     p <- ecs.proxy('plot')
@@ -820,10 +889,11 @@ ec.global <- function(options = NULL) {
 #'
 #' @param wt An \code{echarty} widget as returned by [ec.init]
 #' @param name Name of existing theme file (without extension), or name of custom theme defined in \code{code}.
-#' @param code Custom theme code in JSON format, default NULL.
+#' @param code Custom theme as JSON formatted string, default NULL.
 #' @return An \code{echarty} widget.
 #'
-#' @details Just a few themes are included in folder \code{inst/themes}. The entire collection could be found \href{https://github.com/apache/echarts/tree/master/theme}{here} and copied if needed.
+#' @details Just a few built-in themes are included in folder \code{inst/themes}. The entire collection could be found \href{https://github.com/apache/echarts/tree/master/theme}{here} and copied if needed.\cr
+#'   To create custom themes or view predefined ones, visit \href{https://echarts.apache.org/en/theme-builder.html}{this site}.
 #'
 #' @examples
 #' mtcars %>% ec.init() %>% ec.theme('dark-mushroom')